Version 0 of Newslister
Create a Python package containing a library and scripts for downloading
Google News results and aggregating the results. The library provides the
ability to bring up a set of SOCKS5 proxies on AWS, parse GoogleNews
webpages, and output the aggregation results to a spreadsheet. Scripts
are provided wrapping this behaviour.